Remove the interface/implementation split from DocumentService
